requirements.txt
# Core LLM + API

openai>=1.30.0

# LangChain ecosystem (LCEL + agents + tools)

langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

# LCEL + Runnables support

langsmith>=0.1.0

# Tavily search tool

tavily-python>=0.3.0

# Web scraping

beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# Async support (important for agents + tools)

httpx>=0.27.0
aiohttp>=3.9.0

# Environment variable management

python-dotenv>=1.0.0

# Data handling (optional but useful for research pipelines)

pandas>=2.2.0

# JSON + parsing helpers

orjson>=3.10.0

# Logging / debugging

rich>=13.7.0

# Token counting (useful for agent cost control)

tiktoken>=0.7.0

streamlit
